+++ server/src/main/java/wap/web2/server/controller/UserController.java
@@ -1,10 +1,10 @@
 package wap.web2.server.controller;
 
 import wap.web2.server.exception.ResourceNotFoundException;
-import wap.web2.server.model.User;
+import wap.web2.server.domain.User;
 import wap.web2.server.repository.UserRepository;
-import wap.web2.server.security.CurrentUser;
-import wap.web2.server.security.UserPrincipal;
+import wap.web2.server.ouath2.security.CurrentUser;
+import wap.web2.server.ouath2.security.UserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +16,7 @@ public class UserController {
     @Autowired
     private UserRepository userRepository;
 
+    // 테스트용, 추후에 수정
     @GetMapping("/user/me")
     @PreAuthorize("hasRole('USER')")
     public User getCurrentUser(@CurrentUser UserPrincipal userPrincipal) {
