+++ server/src/main/java/wap/web2/server/domain/AuthProvider.java
@@ -1,4 +1,4 @@
-package wap.web2.server.model;
+package wap.web2.server.domain;
 
 public enum  AuthProvider {
     local,
