+++ server/src/main/java/wap/web2/server/ouath2/config/SecurityConfig.java
@@ -1,35 +1,24 @@
-package wap.web2.server.config;
+package wap.web2.server.ouath2.config;
 
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.JWKSet;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
-import com.nimbusds.jose.jwk.source.JWKSource;
-import wap.web2.server.security.CustomUserDetailsService;
-import wap.web2.server.security.RestAuthenticationEntryPoint;
-import wap.web2.server.security.TokenAuthenticationFilter;
-import wap.web2.server.security.oauth2.CustomOAuth2UserService;
-import wap.web2.server.security.oauth2.HttpCookieOAuth2AuthorizationRequestRepository;
-import wap.web2.server.security.oauth2.OAuth2AuthenticationFailureHandler;
-import wap.web2.server.security.oauth2.OAuth2AuthenticationSuccessHandler;
+import wap.web2.server.ouath2.security.CustomUserDetailsService;
+import wap.web2.server.ouath2.security.RestAuthenticationEntryPoint;
+import wap.web2.server.ouath2.security.TokenAuthenticationFilter;
+import wap.web2.server.ouath2.security.oauth2.CustomOAuth2UserService;
+import wap.web2.server.ouath2.security.oauth2.HttpCookieOAuth2AuthorizationRequestRepository;
+import wap.web2.server.ouath2.security.oauth2.OAuth2AuthenticationFailureHandler;
+import wap.web2.server.ouath2.security.oauth2.OAuth2AuthenticationSuccessHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.BeanIds;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.JwtEncoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
