+++ server/src/main/java/wap/web2/server/ouath2/security/TokenAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package wap.web2.server.security;
+package wap.web2.server.ouath2.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -13,7 +13,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
-import wap.web2.server.security.TokenProvider;
 
 import java.io.IOException;
 
@@ -23,7 +22,7 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
     private TokenProvider tokenProvider;
 
     @Autowired
-    private wap.web2.server.security.CustomUserDetailsService customUserDetailsService;
+    private CustomUserDetailsService customUserDetailsService;
 
     private static final Logger logger = LoggerFactory.getLogger(TokenAuthenticationFilter.class);
