+++ server/src/main/java/wap/web2/server/ouath2/security/oauth2/CustomOAuth2UserService.java
@@ -1,12 +1,12 @@
-package wap.web2.server.security.oauth2;
+package wap.web2.server.ouath2.security.oauth2;
 
 import wap.web2.server.exception.OAuth2AuthenticationProcessingException;
-import wap.web2.server.model.AuthProvider;
-import wap.web2.server.model.User;
+import wap.web2.server.domain.AuthProvider;
+import wap.web2.server.domain.User;
+import wap.web2.server.ouath2.security.oauth2.user.OAuth2UserInfo;
+import wap.web2.server.ouath2.security.oauth2.user.OAuth2UserInfoFactory;
 import wap.web2.server.repository.UserRepository;
-import wap.web2.server.security.UserPrincipal;
-import wap.web2.server.security.oauth2.user.OAuth2UserInfo;
-import wap.web2.server.security.oauth2.user.OAuth2UserInfoFactory;
+import wap.web2.server.ouath2.security.UserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.core.AuthenticationException;
